Feat: 컬럼 정렬 기능 및 화면 추가

+++ src/main/java/com/example/collaborativetools/board/dto/GetDetailResponseBoardDTO.java
@@ -8,7 +8,7 @@ import lombok.Builder;
 
 @Builder
 public record GetDetailResponseBoardDTO(
-        Long columnId, String columnTitle, Integer sequence, List<CardInColumnDTO> cardTitleList
+        Long columnId, String columnTitle, Double sequence, List<CardInColumnDTO> cardTitleList
 ) {
 
     public static GetDetailResponseBoardDTO of(Columns columns, List<CardInColumnDTO> cardDTOList) {
