+++ src/main/java/com/example/collaborativetools/column/dto/request/ColumnUpdateRequest.java
@@ -17,7 +17,7 @@ public class ColumnUpdateRequest {
 	private final String title;
 
 	@NotNull
-	private final Integer sequence;
+	private final Double sequence;
 
 	public Columns toEntity(Board entity) {
 		return Columns.create(title, sequence, entity);
